refactor(cli): organize environment and secret commands
