+++ app/src/cli/commands/environment/handler.rs
@@ -1,9 +1,9 @@
-use crate::cli::{output, prompt};
+use super::EnvCommand;
 use crate::cli::{
-  args::EnvCommand,
   client::{self, ApiClient},
   local_config,
 };
+use crate::cli::{output, prompt};
 use crate::constants::api as api_paths;
 use anyhow::{Context, Result};
 use reqwest::Method;
