+++ app/src/cli/commands/environment/mod.rs
@@ -1,3 +1,6 @@
+mod args;
 mod handler;
 
+pub use args::EnvCommand;
+pub(crate) use args::HELP;
 pub(super) use handler::{env_id, execute, resolve_environment};
