+++ app/src/cli/commands/secret/mod.rs
@@ -1,3 +1,6 @@
+mod args;
 mod handler;
 
+pub(crate) use args::HELP;
+pub use args::SecretCommand;
 pub(super) use handler::execute;
