+++ app/src/constants/help.rs
@@ -26,86 +26,6 @@ Common server options:
 Run 'dopbase help <command>' for details on any command.
 ";
 
-pub(crate) const ENV_HELP: &str = "\
-Examples:
-  dopbase env create payment-service/production
-  dopbase env list payment-service
-  dopbase env show payment-service/production
-  dopbase env default payment-service/development
-  dopbase env rename payment-service/production prod
-  dopbase env delete payment-service/staging
-";
-
-pub(crate) const ENV_DEFAULT_HELP: &str = "\
-Examples:
-  dopbase env default payment-service/development
-  dopbase env default --clear
-";
-pub(crate) const ENV_CREATE_HELP: &str =
-  "Examples:\n  dopbase env create payment-service/production\n";
-pub(crate) const ENV_LIST_HELP: &str = "\
-Examples:
-  dopbase env list
-  dopbase env list payment-service
-";
-pub(crate) const ENV_SHOW_HELP: &str = "\
-Examples:
-  dopbase env show payment-service/production
-  dopbase env show env_482731
-";
-pub(crate) const ENV_RENAME_HELP: &str =
-  "Examples:\n  dopbase env rename payment-service/production prod\n";
-pub(crate) const ENV_DELETE_HELP: &str = "\
-Examples:
-  dopbase env delete payment-service/staging
-  dopbase env delete payment-service/staging --yes
-";
-
-pub(crate) const SECRET_HELP: &str = "\
-Examples:
-  dopbase secret list payment-service/production
-  dopbase secret set payment-service/production API_KEY
-  dopbase secret get payment-service/production API_KEY
-  dopbase secret get payment-service/production API_KEY --reveal
-  dopbase secret delete payment-service/production API_KEY
-
-Use PROJECT_REF/ENVIRONMENT_NAME for readable references. PROJECT_REF can be a
-project ID or name. Immutable environment IDs such as env_482731 also work.
-";
-
-pub(crate) const SECRET_LIST_HELP: &str = "\
-Examples:
-  dopbase secret list payment-service/production
-  dopbase secret list env_482731
-
-Run `dopbase env list` to find an environment.
-";
-
-pub(crate) const SECRET_SET_HELP: &str = "\
-Examples:
-  dopbase secret set payment-service/production API_KEY
-  dopbase secret set payment-service/production API_KEY --stdin
-  printf '%s' \"$API_KEY\" | dopbase secret set payment-service/production API_KEY --stdin
-
-Without --stdin, Dopbase uses a masked prompt and displays * for each character.
-
-With --stdin, Dopbase reads the value until EOF. In a terminal, paste or type
-the value, then press Ctrl+D. On Windows, press Ctrl+Z, then Enter. Piped input
-is read exactly as supplied.
-";
-
-pub(crate) const SECRET_GET_HELP: &str = "\
-Examples:
-  dopbase secret get payment-service/production API_KEY
-  dopbase secret get payment-service/production API_KEY --reveal
-";
-
-pub(crate) const SECRET_DELETE_HELP: &str = "\
-Examples:
-  dopbase secret delete payment-service/production API_KEY
-  dopbase secret delete payment-service/production API_KEY --yes
-";
-
 pub(crate) const IMPORT_HELP: &str = "\
 Examples:
   dopbase import payment-service/production .env.production
